Base product cell implemented.
