+++ N11Case/Screens/Products/View/Cells/NormalProductCell.swift
@@ -10,75 +10,7 @@ import NetworkProvider
 import SDWebImage
 import UIKit
 
-
-protocol ConfigurableCell {
-    associatedtype DataType
-    func configure(with data: DataType)
-}
-
-class NormalProductCell: UICollectionViewCell, ConfigurableCell {
-
-    @IBOutlet weak var rateView: CosmosView!
-    @IBOutlet weak var productImageView: UIImageView!
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var priceLabel: UILabel!
-    @IBOutlet weak var discountPriceLabel: UILabel!
-    @IBOutlet weak var sellerLabel: UILabel!
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        // Reset visibility for stack views and labels
-        titleLabel.isHidden = false
-        priceLabel.isHidden = false
-        discountPriceLabel.isHidden = false // Discount label is hidden by default
-        productImageView.isHidden = false
-        rateView.isHidden = true
-        discountPriceLabel.isHidden = true
-
-        // Reset label texts
-        titleLabel.text = nil
-        priceLabel.text = nil
-        discountPriceLabel.text = nil
-        sellerLabel.text = nil
-
-        // Reset image view
-        productImageView.image = nil
-        productImageView.sd_cancelCurrentImageLoad() // Cancel any ongoing image downloads
-
-        // Reset additional properties if any
-        priceLabel.attributedText = nil
-        priceLabel.textColor = UIColor.productDiscountPrice // Reset to default text color
-    }
-
-    func configure(with product: Product) {
-        titleLabel.text = product.title
-        priceLabel.text = CurrencyFormatter.turkey.format(price: product.price)
-        sellerLabel.text = product.sellerName
-        rateView.settings.starSize = 15
-
-        if let discountedPrice = product.instantDiscountPrice {
-            discountPriceLabel.isHidden = false
-            priceLabel.strikeThrough(true)
-            discountPriceLabel.text = CurrencyFormatter.turkey.format(price: discountedPrice)
-        } else {
-            priceLabel.strikeThrough(false)
-            priceLabel.textColor = UIColor.productNormalPrice
-        }
-
-        if let rate = product.rate {
-            rateView.rating = rate
-            rateView.isHidden = false
-        }
-
-        if !product.image.isEmpty {
-            productImageView.downloadImage(url: product.image)
-        }
-    }
+class NormalProductCell: BaseProductCell {
 }
 
 extension UIImageView{
